apps/lead_agent/requirements.txt
# Core
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
psycopg2-binary==2.9.10

# HTTP / scraping
httpx==0.28.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Azure
azure-storage-blob==12.24.0
azure-servicebus==7.12.3
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0

# Utilities
tenacity==9.0.0
structlog==24.4.0
python-dotenv==1.0.1
click==8.1.8
